Clicking an error in the list focuses the cursor on it in your code
Notably, we do *not* use state/props to tell the editor to do this.
Instead, we grab the `Editor` component directly and call a method on it
to jump to a particular position in the document.

This is because the cursor position is not a piece of information we
care about in general. It wouldn’t make sense as a piece of state or a
property, because then we’d have to keep it up to date all the time as
the user moved around the document; this work would bring no actual
benefit.

The desire to move to the source of a particular error is entirely
transient. The error does not become “selected” or anything like that.
We simply want to make a one-off motion within the editor to react to
the click.

Maybe I’m not thinking reactively enough here. But my instinct is that,
in this case, it makes sense to do this imperatively rather than through
the usual reactive flow of information.