Create project in ProjectStore
Also moves defaults from `config` module into `ProjectStore` and removes
the need to deep-clone the object.

It does feel a bit weird to add a new property to the *action* during
this part of the dispatch cycle, but I couldn’t figure out a better way
to make the information of “which project just got created” to the
`CurrentProjectStore`.