Output component holds current project key in state
Otherwise the other state becomes derived state, which introduces the
possibility for things to get out of sync (a real problem here that only
does not cause a bug because of coincidental facts about the change
notification cycle of the error store).

Note that `calculateState` becomes a local function rather than a method
on the component—I believe this is a general pattern, and having a
`_calculateState` is an anti-pattern, because calculating the state
should never require reference to anything else about the component
(particularly the props).