Components either manage own state, or they don’t
Make a clear distinction between View Controllers and managed Views.
Just like the documentation says.

In particular: refactor all components that refer to props when
calculating state. This violates the [Props in getInitialState
antipattern](https://facebook.github.io/react/tips/props-in-getInitialState-as-anti-pattern.html)
and in each instance leads to specific conceivable bad behavior
(although none of this behavior was evident because of specific details
of how the dispatch->update lifecycle played out).

A pattern that emerges is that the state calculation in view controllers
is done by a local function rather than a private method of the
component. There is no reason that state calculation should need to
refer to properties of the component, since this inevitably means that
the state is derived from the props.