Make runtime error detection work in Safari
Safari considers `srcdoc` iframes to be from a different origin, and
does not report line numbers. So, if the browser is Safari, we instead
fall back to the `contentDocument.write` method of populating the
iframe, in which Safari does accurately report line numbers.

However in this case (for some reason) the line numbers are local to the
contents of the script tag, rather than (as is typically the case for
iframes) relative to the entire file. So, we also calculate the line
number offset differently.