Less terrible layout for toolbar
Absolutely position the toolbar and make space for it using padding.
Means that the whole thing now takes up 100% of the viewport height,
instead of more than 100%. Also means that opening the toolbar covers
the environment rather than pushing it down, which I think is better.

Ultimately I think flexbox is probably the right move here but this is
an improvement from the previous layout.