Ensure the existence of <html>, <head> and <body> elements
When rendering preview, don’t assume that `document.documentElement`,
`document.head`, and `document.body` exist—if the HTML input to the
preview is malformed, they may not. So, check to see if they exist, and
create them if not.

“But why would we be rendering a preview with malformed HTML?”, you may
wonder. As it happens, any time the user types into the editor, the
error state of the project is reset. Then validations are run, and then
the error state is flipped back on based on validations. This is not
great, but best tackled separately.