Make sure to save the current project state when appropriate
Which is always—any time we are saving a project, it is known to be the
current project, so be sure to update that key as well. That way we can
be more discerning about when we save the current project key (e.g., not
when we create a new project, until we start editing it).