Import from gist
When the `Workspace` bootstraps, it checks the query string for a `gist`
parameter. If it finds one, this is passed to the `bootstrap` action,
which will load the gist as the current project if an ID is present.

Also adds a nice bit of logic to the `projects` reducer which ensures we
won’t hold on to an untouched project when switching away from it.