Stub out SVGs in test environment
Since tests run in a normal Node environment, you can’t require SVGs.
This does call into question the wisdom of the whole requiring-of-SVG
thing in the first place, but alas, it’s quite convenient.

Deal with the problem by creating a single SVG utility that exports
useful SVGs—uses CommonJS `require` to conditionally load in SVGs except
in the test environment; in test it instead exports a stub React
component.