Don’t include libraries in config; run bower postinstall
No need to run as part of gulp; it’s part of the install process, so run
it as an NPM postinstall hook.

Also, it should be possible to require in `config` without needing the
libraries to be installed, so move the library config to its own file
(also just better for code organization).