Keep iframe in the DOM
Each time you press a key, the error state immediately becomes
`validating`, which means the preview should no longer render. However,
that means that, previously, every keystroke resulted in the preview
iframe being torn down and then re-added to the DOM. This does not seem
terribly efficient.

Instead, keep the iframe in the DOM at all times; if we’re not in a
valid state to render the preview, simply hide the iframe and set its
source to an empty string.