Resize editor when other editors minimized/maximized
Slightly clunky, but the `Workspace` now includes a `percentageOfHeight`
property that it passes to the editor. The editor doesn’t ever use the
actual value, but it does detect changes to the value, and resizes the
editor itself on such a change.

Uses `requestAnimationFrame` because we need to let React flush to the
DOM before doing the update.