Fix sidebar padding in Firefox
Absolutely positioned elements in Firefox take percentage cues from
their nearest positioned parents. In this case that meant that Firefox
was drawing the padding-bottom of 50% as a percentage of the overall
page.

So, make the `.sidebar-arrow` itself an absolutely positioned container,
and then draw the arrow out of a pseudoelement that is positioned inside
of it.

Confirmed in Chrome, Firefox, and Safari.