Better error for missing tokens
If you write:

```javascript
if (1 === 2)
```

JSHint will generate a `W116` warning, which is the one used for strict
operators among other things. In the general case, this is of the form
“we expected A where you wrote B”, but in this case B will be an empty
string. So, in the case where we get an empty “what you wrote”
parameter, we generate a better error message to the effect of “you’re
missing an A on this line”.

Also, make that error suppress the less valuable “end of input” error;
also add the same suppression to a similar “missing closing brace”
error.