Show help text in empty editors
Show a message saying, e.g., “Type your CSS code here” in the empty
CSS editor pane.

For this to work, had to extract an `EditorContainer` out of the
`Editor` component—this is because `Editor` never updates (state is
manually managed) but we need to show/hide the component based on the
current state. So `EditorContainer` is a normal stateless component
while `Editor` is a managed component.