Distinguish between missing semicolon and extra tokens
This stylesheet:

```css
p {
   padding: 1px 1px 1px 1px 1px;
}
```

Yields an “extra tokens after value” error from PrettyCSS. This is
what we’d expect, since there is an extra `1px` at the end of the value,
which is not a valid way to write `padding`.

However, this will generate the same error:

```css
p {
   padding: 1px
   margin: 4px;
}
```

Since CSS is insensitive to newlines, the `margin` on the second line
appears to be a continuation of the value for `padding` on the first
line.

We can heuristically identify situations where the “extra tokens after
value” error is the result of a missing semicolon:

* The token that fails to parse is the first non-whitespace string on
  its line
* The previous line does not end in a semicolon