Better error message for space after the opening angle bracket of a tag
If you put space right after the opening angle bracket of an HTML tag,
e.g. `< p>`, that’s not valid HTML.

From the parser’s standpoint, this looks like a tag whose name is `""`,
the empty string. However, we can heuristically detect that this is
actually a tag with an erroneous space character before the tag name,
and give a more informative error message.