Set up new action creators for typing
We want to keep track of whether the user is typing (heuristically,
meaning whether they have typed a key in the last second). Set up action
dispatchers and tests to do this. Involves setting up some useful
infrastructure and patterns for testing asynchronous actions, of which
we have a whole bunch.