Add currently typing lifecycle to components
Specifically, the Workspace component dispatches a `userTyped()`
whenever it receives a callback from an `Editor` that the input has
changed; and it uses the `ui.editors.typing` state to determine whether
to show a spinner instead of errors.

This means that we will *only* show a spinner if there are validation
errors *and* the user has recently typed. Importantly, we will no longer
show validation errors if:

* The user has just finished typing, making a previously invalid source
  valid
* Code has just failed validation, but the user was not recently typing
  (e.g. when a project with invalid code is loaded from storage)