Remove delayed error display
If the user is in the middle of typing code, we don’t want to show
validation errors about their half-typed code. So, instead, we show a
spinner if the user is typing *and* there are currently errors in the
source.

Previously, we would do that using the following logic:

* After validation, if there were errors, set a property in the store
  instructing the UI to delay displaying the errors
* Also after validation, invoke a debounced action dispatcher to set the
  error display delay flag back to false

This had a couple of disadvantages:

* A session of typing that ended in valid code would still show the
  spinner for a second before turning green and showing the preview
* When a project with invalid code was loaded from storage, the spinner
  appeared for a second before showing the errors. This is incoherent

So, the code for that has now been removed, as it has been replaced with
better logic.