Handle click of error associated with minimized editor
In this situation:

* Type invalid code into an editor
* Minimize that editor
* Then click on the error message in the list

Expected behavior: The editor is maximized and focused, with the cursor
at the position of the erroneous code.

Previous behavior: uncaught exception.

This pull request afforded an opportunity to finally fix one of my least
favorite things about Popcode’s implementation, namely that the
`Workspace` held references to the `Editor` component instances that it
owned. Worse, when an error was clicked, the `Workspace` would call a
method *directly on the `Editor`*. This is not how components are
supposed to work.

I have had a lot of trouble coming up with a satisfying way to represent
what feels like a completely transient event—“the user clicked an
error”—and a completely transient reaction—“move the cursor in the
appropriate editor to the position of the error”—in Flux state. But I
think this approach is pretty good:

* There is a `ui.editors.requestedFocusedLine` state property. This is
  usually `null`, but when populated, it contains information about a
  line that the user wants the cursor to move to, but which we have not
  yet actually moved the cursor to.
* When the user clicks an error, we dispatch a
  `USER_REQUESTED_FOCUSED_LINE` action, which causes the aforementioned
  property to populate.
* The `Editor` component pays attention to this property; if it receives
  a value that corresponds to its language, it moves the cursor
  appropriately, and then immediately dispatches an
  `EDITOR_FOCUSED_REQUESTED_LINE` action.
* This action causes the `ui.editors.requestedFocusedLine` property go
  back to `null`.

Unsurprisingly, migrating this interaction to a pure one-way data flow
makes it a lot easier to work with and reason about. In general, we now
know that, at any given time, we’re in one of three situations:

1. The user hasn’t clicked an error.
2. The user clicked an error, but we haven’t moved the cursor yet.
3. The user clicked an error, and we moved the cursor.

States 1 and 3 end up looking identical in the state tree.

This approach solves a specific problem with a naïve solution to the
original bug. When an error is clicked, we need to maximize the editor.
This is done by issuing a `COMPONENT_MAXIMIZED` action. However, this
won’t cause the component to mount synchronously; we need to wait for
React to flush component state. That means, if we want to call a method
directly on a component we’re holding a `ref` to, we need to do so in
some kind of next-tick block, which is unsatisfying and fragile.

Instead, by using Flux state to communicate the user’s intent to focus a
line, we stop having to worry about the exact sequence of rendering.
When the `Editor` component mounts, it checks to see if there is a
pending focused-line request, and focuses the line if there is. It also,
of course, checks for focused-line requests when props change. So, the
right thing happens regardless of the order things happen in.