Only include libraries we’ll need
A huge amount of bundle size was going toward loading a bunch of
libraries that there is no known demand for.

For now, we’ll just cut it down to:

* jQuery
* Lodash
* Mustache
* Bootstrap

Once we have asynchronous loading for libraries (less trivial than
setting it up for linters), we can go back to being generous with the
library list.