Don’t bundle useless ACE workers
The ACE modes conditionally `require` in their workers in the case where
workers are enabled, which they never are in Popcode. However, since the
`require`s are statically analyzed during compilation, the unused code
still ends up in the bundle.

Not anymore! The `null-loader` is now in effect for the modules in
`brace/worker`, meaning those `require`s now evaluate to an empty
module.