Run release build tests using BrowserStack
BrowserStack allows us to run tests on a bunch of real browsers. Cool!

One problem is that we need credentials to run tests on BrowserStack.
And since this is an open source project, that either means making my
credentials public, or only allowing them to be used when run from the
canonical `popcodeorg` repo. For now, we’ll go with the latter.

This means that PR tests will only be run in Firefox, which is available
locally on Travis. This is not the end of the world, as we will still
run the tests on the full suite of browsers before cutting a new
release.