Upgrade github-api to 2.4.0
Prior to 2.4.0, certain errors are not handled correctly, resulting in
the reported error being an error in the error handler. This makes it
impossible to know what the original error was.

This was [fixed in 2.4.0](https://github.com/michael/github/pull/354),
meaning we will now get meaningful error reports for errors when
interacting with the Gist API.

Apparently the 2.4.0 build failed, so there is no 2.4.0 on npm. For that
reason we include the dependency off of GitHub, which means `dist` isn't
built, which means we need to compile in the source files. So, a bit of
extra configuration in WebPack to compile ES6 and also to point at the
right module lookup paths.