Prevent rapid clicking of GitHub export button
GitHub export often takes a while, which may inspire some students to
close the window and try again. This risks creating duplicate gists, and
possibly getting the student’s account flagged as a bot.

So, while the export is in progress, turn the “export” button into a
spinner; nothing happens when you click the button in this state.

When the gist export either succeeds or fails, the button will be
restored to a functioning, non-spinner state.