Spinner while gist export is in progress
When the user initiates a gist export, we need to synchronously open the
new window for the gist, because Chrome's pop-up blocker will kick in
if we try to open a window in an execution context that wasn’t initiated
by a user interaction. When gist exports are slow—which is a lot, in
situations where the network is bad—opening a blank window that then
sits there for many seconds makes the whole thing appear broken.

So, given the need to immediately open the window, and also the
impossibility of re-focusing the Popcode window (as far as I can tell),
the best solution is to make the blank window appear less broken. To
that end, the window now just displays a big spinner until we can point
it at the exported gist.