Remove notifyAndRejectApiError from Gist client
With the updated version of the GitHub API client, we now get reasonably
useful error information when something goes wrong. So, there’s no need
to intercept errors and report them to Bugsnag ourselves.

This is actively preferable because these methods should not assume that
errors are uncaught (as in the case of using `loadFromId` to verify that
GitHub is able to serve a gist).