Ensure we can read gist from GitHub API before attempting to open it
Various classes have observed situations in which newly-created gists,
when opened in the new tab, yield a GitHub 404 page. This is odd
behavior, considering we are just using the `html_url` returned by the
GitHub API.

Bit of a shot in the dark here, since the error isn’t readily
reproduced, but my theory right now is that we’re hitting an eventual
consistency situation or a slightly laggy database replica on GitHub’s
end.

To try to combat this, after creating (and possibly updating) the gist,
we send an API request to fetch the newly created gist. If the request
fails, we’ll retry with exponential backoff several times.

The hope is that a 404 on the web URL will correlate with a 404 in the
API. There is certainly no guarantee of this, but it’s worth a shot.