Ensure Pop stays within bounds
Previously we had an issue with Pop encroaching into the button area of
the dashboard with very small viewport heights. This fixes that issue.

In Chrome, the behavior is ideal: Pop will resize to accommodate both
the width and height available to it.

In Firefox and Safari, Pop does not resize to accommodate small vertical
space, but its container is `overflow:hidden` so at least it will not
encroach.

This is accomplished by removing the absolute positioning of Pop within
an expanding flex container. Instead, we add a expanding flex spacer
above Pop, and then set Pop to take up only the space it needs. We
further set a `max-width` and `max-height` on the SVG itself.