Don’t allow omission of any close tags
In HTML5, a few close tags can be omitted, e.g. `<p>`. So, this is valid
HTML:

```html
<!doctype html>
<html>
  <body>
    <p>My paragraph
  </body>
</html>
```

However this confusing, inconsistent, and bad practice. Not something we
want to allow students to do.

So, we override the Slowparse parser to clear the list of tags that may
be omitted. Directly reaching into the parser’s prototype and modifying
one of its properties isn’t the most elegant-feeling solution, but it’s
the only way I could find to do it given the API that Slowparse
currently exposes.