Use location.href workaround to work around Chrome closing bug in popout
It appears that a recent version of Chrome has introduced a regression
that causes newly-opened windows to immediately close if they are opened
with a data URI. This behavior definitely used to work fine, and it also
works as expected in Firefox.

Even more oddly, I do not see this behavior in development, only
production. Others have seen it in development, though.

Happily, opening `about:blank` and then immediately setting the
`location.href` of the new window to the desired location (data URI)
seems to work fine.