Load preview in a sandboxed iframe
If we run code in a same-origin iframe (as we have been previously),
JavaScript running in that frame can access `window.top`. I’m not
specifically sure how one can get from `window.top` to sensitive data,
although `webpackJsonp` is defined globally, which I’m guessing can be
used to load modules. From there, it would be fairly trivial to get
access to a GitHub access token and other private information.

Given that a simple gist link could be used to distribute malicious
code, this situation is obviously not tenable. So, switch to using a
frame with a `srcdoc` and the `sandbox` attribute configured
appropriately. Most to the point, it does *not* have `allow-same-origin`
set, so the code runs in an untrusted cross-origin iframe, which
prevents it from accessing `window.parent`, `window.top`, etc.

This has a couple of disadvantages, namely:

* We have to drop support for IE and Edge, since neither supports the
  `srcdoc` attribute.
* Safari, when running code in a sandboxed iframe, does not give
  meaningful information to the `onerror` callback, so we can’t capture
  line information. So, instead, we just hard-code all errors in Safari
  to come from line 1.