Use postMessage for infinite loops
Previous implementation of loop protection relied on injecting the
application’s instance of `loopProtect` directly into the iframe. No
longer possible, so instead we load `loopProtect` as a preview frame
library, and set it up to send a `postMessage` to the parent window if
there is a detected infinite loop.