Retry all network requests to GitHub
When exporting a gist, any of the network requests may fail in
unreliable network environments. So, in the event of a network error, we
want to retry the operation several times:

* API request to create gist
* API request to update gist with URL
* API request to load gist by ID (for gist import)

I also took out the code that attempts an API request to load the gist
after the export completes, because there is no evidence that this helps
prevent a 404 when redirecting to the gist, and it’s another network
request that can be slow or fail.