Browser-friendly stylelint build
The default `stylelint` module wants to load configuration files from
the filesystem in a dynamic way that isn’t possible with `brfs`. We
don’t actually need the file-loading behavior, so instead, we hand-roll
a `stylelint` object in the same manner as
[`createStylelint`](https://github.com/stylelint/stylelint/blob/7.5.0/src/createStylelint.js),
but using configuration defined in code rather than the filesystem. This
definitely builds in knowledge of some non-public APIs in `stylelint`,
but I can’t think of a better approach.

Also a decent amount of Webpack wrangling to get the modules we do
import from stylelint to build. In particular, they *still* want to hit
the filesystem, although they don’t actually need to in order to work
for us. So, instead of building those modules with `brfs`, just use the
string replace loader to turn `require("fs")` into a null object; use
the `substitute` loader to stub out the `isAutoprefixable` method; and
use the null module loader to stub out rules that want access to a
gigantic database of browser capabilities. The usage of
string-replacement is a fragile hack, but I can’t think of another way
to *only* stub out `fs` in that particular context.