Turn off eslint camelcase property enforcement
Unfortunately, to use the `classnames` library to dynamically generate a
set of DOM classes based on boolean tests, we need to construct objects
whose keys are underscored properties. Since eslint allows no nuance in
whether we enforce camelcasing of object properties, the only option is
to just disable this rule.