More reliable location of error tokens
Found a case in which the Esprima validator was unable to identify a
token at the location of a parse error. Upon further investigation, it
turned out that in this case the `row` and `column` reported in the
error identified a location in the source that did not actually exist.
However, it’s possible to have the tokenizer output tokens with the
index of the token start and end in the overall source, rather than
line/column information; the corresponding `index` reported in the error
object appears to be more reliable.

As a bonus, the logic for locating the token is a bit simpler.

So, switch to using index rather than line/column to match up the error
location to a token.