Purge CloudFlare cache after deployment
CloudFlare revalidates is cached copy of Popcode assets fairly
regularly, but after a deploy we want to immediately purge the cache so
that an old version of the code isn’t served. The main motivation here
is to make sure that the deploy takes immediate effect; however, in rare
circumstances, a mismatch between the cached version of different files
might lead to bad behavior. So, ensure that after a deploy, we’re
serving the most up-to-date version of every file.