Flatten user data structure in store
Until now, we’ve been directly storing the profile information in the
`github` key of the Firebase response in the `info` key of our own user
representation in the store. This means that knowledge of the particular
structure of the Firebase user representation leaks throughout the
application.

Now, instead, the user reducer maps the Firebase structure on to a
normalized, flat user representation for internal use. Most relevant
information is now a key in the top-level user object; the only
exception is `accessTokens`, which is a map of provider to access token,
in the unlikely(?) event we’ll be supporting authentication with other
providers in the future.

This is in preparation for the Firebase SDK upgrade, which will in fact
give us a different structure of the Firebase user representation.

Also reads the `uid` from the top-level Firebase response rather than
the `auth.uid` key, because this seems more correct (it’s a top-level
fact about the Firebase user).