Upgrade to Firebase 3 SDK
Upgrades our `firebase` dependency to version 3 of the SDK, and our
integration and specs to conform to the new API. The application
basically works in this state, except for restoring a session at
bootstrap time, which does not work properly because Firebase no longer
provides third-party credentials (e.g. GitHub) when restoring a session.

Unlike Firebase 2, Firebase 3 passes a `credential` object separate from
the `user` object containing the user profile after a successful OAuth
login. So, modify the `logIn` action dispatcher to take two arguments,
and the reducer to handle them accordingly.

This slightly changes the schema of the `user` in the store: namely, the
`accessTokens` map now uses the key `'github.com'` rather than just
`'github'`. There is also no concept of a global `provider` of the user
(this concept has been removed from the Firebase user object, and never
made a ton of sense anyway, since it is a fact only about the current
login session, not about the user themselves). Modify the `Gists` client
to account for these changes.

One thing to note is that `onAuthStateChanged()` does not yield any
credential information, meaning we will need to store credentials in
session storage ourselves to access them after the page has reloaded.
This is not done yet, and tests do not pass because currently the call
to `userAuthenticated` doesn’t pass a `credentials` object at all.

We also update the configs to point to a new development instance of
Firebase. Since the callback URL has changed, it’s easier to set up a
new Firebase app with a new OAuth integration on GitHub, rather than
trying to switch the configs back and forth a whole bunch.

Finally, change the approach to async testing, using a setTimeout to
allow asynchronous operations to complete.  `MockFirebase` needs to
yield asynchronously to the `onAuthStateChange()` method, because real
Firebase will always do so, and the implementation will only work if
`offAuth` is defined before it is used. So, we use sinon’s `yieldAsync`,
but this method yields using a `setTimeout` rather than a promise
resolution, so allowing promise ticks to run is insufficient.

Instead, replace `promiseTicks` with `waitForAsync`, which returns a
`Promise` that resolves after an environment-minimum timeout fires. This
does not slow down test suite execution appreciably.